fix(MCU_PORT/timer_general): 🐛 fix interrupt freezing after while
The problem arises when the software timer stop and start the oc timer
after the base timer start working, this was leading to missing the
interrupt because the base timer counter value exceeds the OC timer
pulses. So instead of arming the OC timer with 0 counter we starting it
with the base timer counter and add the period we want the OC timer to
fire on.